requirements.txt
Django>=2.2.6
djangorestframework>=3.9.4
pytz>=2019.3
sqlparse>=0.3.0
uwsgi>=2.0,<2.1
dj-database-url>=0.5,<0.6
psycopg2-binary>=2.8.4 
django-filter
drf-nested-routers
